Fix bug which caused anchors to be stripped from cached executables
Summary:
The InputInfos and OutputInfos entries in PoplarExecutableInfo, which were previously used
to export the PopEF anchors, are not reloaded from saved executables, so anchors were
being stripped when exports were done when using cached executables. This patch changes
the export process to extract the anchor information from sources which are reloaded,
so that anchors are no longer stripped.

TF2.4 Only

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, anthonyb, jackh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, anthonyb, jackh

Subscribers: jackh, georgep

Maniphest Tasks: T48681

Differential Revision: https://phabricator.sourcevertex.net/D55690