Account for 3D average pool where one dimension is not reduced over
Summary: Currently, a 3D average pool will fail if one dimension is not reduced over because incorrect parameters are passed to Poplibs. There is code that accounts for this possibility in the 2D case, which is extended here to the ND case.

Test Plan: Only the file containing the new tests has been tested, I have not yet run a more general test suite

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T53108

Differential Revision: https://phabricator.sourcevertex.net/D58408