Using vector<Tensor> overload of gcl::allGather
Summary:
Updating all_gather to use gcl::allGather overload which accepts a vector of inputs.
This simplifies the logic in the backend as we were previously packing together compatible inputs, then unpacking the outputs.
This change is also reflected in the python interface which has been updated to optionally accept a list of inputs.

TF1.15 Only

V2 diff: D58356

Identical to V2 diff, just a difference in the surrounding entries in the BUILD file.

Test Plan:
Added 3 tests:
 - Check output when passing a single input
 - Check output when using multiple inputs
 - Check output when using multiple inputs with different dtypes

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk, georgep

Maniphest Tasks: T48294

Differential Revision: https://phabricator.sourcevertex.net/D58431