Executing keras pipeline stages in order
Summary:
TF2.4 Only

Iterating over pipeline stages in id-order when building the graph in keras.
Previously they were built in whatever order the dictionary was in. This caused a bug where we encounter the final pipeline stage but not all outputs have been created.

Test Plan: Added a test which failed before the change, but passes now.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgew

Maniphest Tasks: T50789

Differential Revision: https://phabricator.sourcevertex.net/D58631