Warn about usage of BatchNormalization in pipelined Keras models
Summary:
keras.BatchNormalization layers currently do not properly update their moving statistics during training when inside a pipeline, as they need to update per-batch rather than at the end of the pipeline in the WU. This diff adds a warning to warn users of this when they use a BN layer in a pipelined Keras model.

Fixes T46696
TF2.4 Only.

Test Plan: Added test that checks warning is emitted and checks it is justified by making sure the BN is behaving badly.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T46696

Differential Revision: https://phabricator.sourcevertex.net/D58819