Documenting how to pipeline Model subclasses
Summary:
Documenting how to pipeline subclassed models, with 2 examples.
Adapted from the docs for piplining functional models, with different examples and some additional information which is only relevant to Model subclasses.

TF2.5 Only

Test Plan: The examples are added as tests to so we don't break or forget to update them.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Subscribers: jamiep

Maniphest Tasks: T50789

Differential Revision: https://phabricator.sourcevertex.net/D61053