Deduplicate anchor names when exporting executables to PopEF
Summary:
TensorFlow sometimes produces a bunch of inputs called "dummy_input"
(this comes from tensorflow/core/common_runtime/eager/attr_builder.cc,
and is probably eager mode-specific, but it's not clear what these
inputs are). PopEF uses the names of anchors to match them to feeds
and tensors, so duplicate names are a problem. This patch deduplicates
the names with a numeric suffix to ensure that this constraint is not violated.

TF2.4 Only

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, anthonyb, jackh, gauthamg

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, anthonyb, jackh, gauthamg

Subscribers: gauthamg, jackh, georgep

Maniphest Tasks: T54727

Differential Revision: https://phabricator.sourcevertex.net/D56333