Add code to detect while loop broadcasts
Summary:
This is the start of the code to remove tensor list broadcasts.
This is just the detection phases where it finds inefficient
broadcasts that are only used by slice instructions. The remove function
is blank.

Note this is not valid detection yet as it does not check that the slice index
is less than the indexs updated into.

Test Plan:
included 2 but will focus most testing in diff that removes the instructions as
both need to work together

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, babakk, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, babakk

Maniphest Tasks: T50781

Differential Revision: https://phabricator.sourcevertex.net/D58541