Adding function for replacing dynamic-slice with a multi-slice
Summary: This change adds `TryReplaceDynamicWithMultiSlice` which will replace a given dynamic-slice with a multi-slice if it's 1D slice of size 1. We have this limitation since the change is targeting rnns and it looks like all the dynamic-slices created by test_rnn.py are 1d and of size 1.

Test Plan: New tests.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, vladimirm

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Maniphest Tasks: T54306

Differential Revision: https://phabricator.sourcevertex.net/D60287