Allow slice index analysis to check some unknowns
Summary:
Do a small amount of analysis on the index if it is unknown.
Analyse unknown indexes by keeping track of positive and
negative numbers in a computation. Then iterate through the
non trivial uses of the index from the writes and if it hits the
read we are interested in then we are safe

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk, vladimirm

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm

Maniphest Tasks: T50781

Differential Revision: https://phabricator.sourcevertex.net/D60471