New ver of ApplicationRuntime op accepting filename as Tensor
Summary:
Embedded runtime needs a path, where the model compiled is stored, to
load and run it. As the location of the executable may be known only
during execution, embedded runtime needs to accept it's path in form
of tensor updated on savedmodel load (see Asset).

TF2.5 only

Ref T51881

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh

Subscribers: jakeh, georgep, jacekr

Maniphest Tasks: T51881

Differential Revision: https://phabricator.sourcevertex.net/D59833