Replace uses of inefficient broadcasts with hoisted instructions
Summary:
Once we have detected an in-efficient broadcast, remove it by replacing
all it's uses with the hoisted constant. We then propogate the new
shape through the module. We don't remove the slice at the end, we
just replace it's input in such a way the algebraic simplifier can
eliminate it later for us

`TF1.15 only`

remove logging

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk, vladimirm, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm

Maniphest Tasks: T50781

Differential Revision: https://phabricator.sourcevertex.net/D60836