Add pipelining support for subclassed models
Summary:
Supporting pipelining for subclassed models.

When you manually call build on a subclassed model, the call function is executed with keras.Input tensors (with the specified shape) in order to trace the graph so we can later generate a post order of the layers to use in pipelining. The tracing is done using functions from the functional model which are attached onto the subclassed model in the extension constructor. The build function is overridden using an extension delegate in order to make this work. The tracing does not happen if the build is happening implicitly as part of a call function (if we are in a call context). The graph is only traced when you manually call build, or if you call fit/evaluate/predict without having manually called build. This avoids unnecessary work in simpler applications of the subclassed model where there is no pipelining.

One limitation of using build to trace the graph is that by default you can't pass in any dtype information. To get around this I've made the build function support `keras.Input`s as these can have dtype information. These are what the overridden build function generates internally anyway, and are how the input dtypes for other model types are specified.

Only the layers of the graph, deduced from tracing it, are executed during pipelining. No side effects from the call function (e.g. prints or custom metrics set in call) will occur. The traced graph also persists through saving and loading, utilizing functional model methods.

By default, when loading a subclassed model, we attempt to construct it by calling a parameterless constructor. If this isn't supported on the model then it needs to override from_config to specify how to restore the model. There are relevant error messages to explain this.

Declaring pipeline stages works the same way as with functional models. You can either apply them at layer/node creation with a PipelineStage scope, or assign them later using an api identical to the other models.

If you want to setup pipeline stages using get_pipeline_stage_assignment and set_pipeline_stage_assignment, you need to call build to trace the model before using those functions. If your model already has pipeline stages assigned using PipelineStage scopes, then you don't need to call build manually. fit/predict/evaluate have been changed to call a new _build_with_dtypes function implcitly. This replaces the shapes and dtypes parameters previously on _get_pipline_post_order which were used to make sure sequential models with no input layers are traced. The tracing logic from sequential _get_pipline_post_order has been moved into its implementation of _build_with_dtypes.

I made the decision to only trace in _build_with_dtypes if the graph had not been traced at all. Otherwise, we risk invalidating the users pipeline stage assignments if the trace comes out differently. Instead, it's up to the user to call build between calls to fit/evaluate/predict if they want to change the graph.

There is also a small change to the general extension behaviour. When an extension is applied to an instance, the constructor of the patched class we generate is now changed to additionally call the extension constructor. The only time this makes a difference is if we try to construct a new instance of that patched class, for example if we call model.from_config on the patched instance. Previously we were calling the extension constructors in deserialize_from_config to get around them not being called, but that led to the constructor being called twice in lots of other cases.

TF2.5 Only

Test Plan: Similar tests to other model types but they have been split up more, and a few subclassed-model-specific tests have been added.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jackh, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jackh, samuelh

Subscribers: christiana, alfiee

Maniphest Tasks: T50789

Differential Revision: https://phabricator.sourcevertex.net/D58269