New ver of ApplicationRuntime op accepting filename as Tensor
Summary:
Embedded runtime needs a path, where the model compiled is stored, to
load and run it. As the location of the executable may be known only
during execution, embedded runtime needs to accept it's path in form
of tensor updated on savedmodel load (see Asset).

TF1.15 only

Ref T51881

This is port of https://phabricator.sourcevertex.net/D59833

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T51881

Differential Revision: https://phabricator.sourcevertex.net/D61131