Handling case where we allocate through a reduce
Summary:
Previously, in the rare case we allocate through a reduce we would get a crash.
We would only allocate through a reduce when the shape of the tensor being allocated happened to match the output shape of the reduce.
I added a test to recreate these conditions.

The crash occured because we didn't consider reduce (an op which affects shape) in pathTransform, leading to the tensor being allocated with the wrong shape.

There is additionally a change to prevent inputs other than input 0 (i.e. input 1, the constant) ever being allocated through reduces since it doesn't make sense to do that.
For input 0 we use the default behaviour. Specifying behaviour for input 0 will be in a future diff.

TF1.15 Only

V2 diff: D61204

Identical to the v2 diff, just a trivial conflict in the BUILD file.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T50967

Differential Revision: https://phabricator.sourcevertex.net/D61249