Fixing error with single dict inputs to pipelined keras model
Summary:
Fixing an error when pipelining keras models where if the dataset iterator returns a single dict, it is inerpreted as kwargs by the pipeline infeed wrapper. This behaviour is consistent with repeat loops. When this happened it caused an error because our pipeline stage functions did not accept kwargs.

This change makes the pipeline stage functions accept kwargs and reinterpret them as args is there are no args.

Note that we will still get an error if their dict uses non-string keys as the infeed wrapper will fail to convert the dict contents to kwargs

TF2.5 Only

Test Plan:
Modified an existing test such that it failed before the fix and passes after it.
Adding predict to the test caused the error as the dataset iterator is only returning x in that case. I also added evaluate to the test for coverage.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T56527

Differential Revision: https://phabricator.sourcevertex.net/D61408