Supporting input dicts with non-string keys in keras pipelining
Summary:
Prevents an error when loops / pipelines try to pass a single dictionary from an infeed as kwargs. This causes an error if all of the keys in the dict are not strings.
This error could be triggered when passing a dict with non-string inputs to a pipelined keras model.
The change in the infeed wrapper for loops was just to be consistent with pipelines. It only affects cases which would previously have resulted in errors anyway.

There is an additional change to the regex in a test. Adding a new test caused a number to sometimes become double digits which changed the whitespace. This wasn't properly accounted for in the original regex.

TF2.5 Only

Test Plan: Added a test which failed before the fix and passes after.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T56527

Differential Revision: https://phabricator.sourcevertex.net/D61481