ModuleFlatten hlo pass tweaks.
Summary:
Linter fixes.
Tweak ModuleFlatten pass to flatten while conditions and condition predicates. This is to help make sure that all multi-slices in a module get flattened, otherwise the EmbeddingPlansPreplanning pass will error. Previously they'd be skipped if in a while condition or used as a predicate in a conditional.

Test Plan: CI and new tests.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T54306

Differential Revision: https://phabricator.sourcevertex.net/D61670