Wrap ipu.outlined_function in tf.function
Summary:
The issue in ticket T53922 is that `ipu.outlined_function` gets `tf.python.eager.backprop._gradient_function` as a grad when using TF2. This hasn't been seen before as all examples of `ipu.outlined_function` use the old session API. The gradient function registered with `TFE_Py_RegisterGradientFunction` returns an instance of `tf.python.eager.backprop._MockOp` which does not have a `graph` attribute as it isn't an op (as the name suggests).

The solution is to wrap `ipu.outlined_function` in `tf.function`.

TF2.5 Only

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk, markf

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk

Maniphest Tasks: T53922

Differential Revision: https://phabricator.sourcevertex.net/D59913