Fixing reduce case in path transform
Summary:
This diff fixes the reduce case in path transform.

The case creates the reduce input tensor using createSliceableTensorFromSlice. The error was that slices have the same rank as the tensor they were sliced from, but reduce outputs don't. I just expanded the tensor to fix this.

This should fix the failing tests in D61620.

Test Plan:
Fixed the existing test for this. The reduce was being optimsed out so the case was never being hit.
I also added in a check that the reduce is not being optimised out.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Maniphest Tasks: T50967

Differential Revision: https://phabricator.sourcevertex.net/D61906