Supporting 1-dimensional rhs in concat into dot optimisation
Summary:
Fixes an error enountered in the concat into dot optimisation.
The optimisation acts on concats passed into dot ops.
It removes the concat and dot, and dots each side of the concat separately, then adds them.

The logic for the concat into dot optimisation is copied directly from the upstream AlgebraicSimplifier so this bug is present there too.
This bug is unlikely to be hit, as it cannot happen when the dot strength optimisation is enabled, and we have it enabled by default. The dot strength optimisation is always done instead if it is enabled.

Test Plan: Added a test which failed without the change, and passes with it.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jackh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jackh

Subscribers: jackh

Maniphest Tasks: T50967

Differential Revision: https://phabricator.sourcevertex.net/D61936