Use exact paths when specifying libs.
Summary: Globs such as "lib*/poplar/libpoplar*" will match against libpoplar_mock.so when tf is built against a poplar sandbox, but this made the resulting whl unuseable with the output of poplar_packaging. Removing the glob and specifying the library names in full fixes this issue.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jakeh, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T56876

Differential Revision: https://phabricator.sourcevertex.net/D62232