Preventing repeated work in allocation finder
Summary:
Optimisation for allocation finder preventing repeated work exploring paths which have already been explored.

A case where this optimisation helps is when two slices of the same tensor get added together.
When looking for allocations for the initial tensor, we would end up exploring the part of the path after the add twice as there are two ways to reach it.

As explained in a comment, the only time we might want to reexplore a path is if we get a prefereable overall path (one with no slices) because of the different beginning. This could lead to a finding a better target.

Test Plan:
This change prevents array_elementwise_ops_test_poplar timing out in D61323.
The test repeatedly slices and adds the same tensor, each time exponentially increasing in the number of paths to explore.
The previous findconsumers default behaviour did not allocate through the adds since their shape did not match the shape of the initial tensor.

Added a new test similar to the case above where a tensor's path repeatedly branches (into 2) and recombines, before reaching a dot. One branch has a slice and the other does not. This means out of the 2^n permutations of the path to the dot only one has no slices. The test checks that allocation finder finds that path without timing out. Before the optimisation this test would have timed out.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, samuelh

Maniphest Tasks: T56392

Differential Revision: https://phabricator.sourcevertex.net/D62048