TF1 - Evaluate external captures as constant expressions when outlining
Summary:
This change updates the semantics of outlining to make externally captured arguments to the outlined function eligible for constant evaluation. This is needed as some ops require a compile time constant input, e.g. transpose. We don't enable this for direct arguments to the outline function because this can reduce the benefit of outlining - since a function will be generated for each different constant argument.

This fixes errors like...

```
tensorflow.python.framework.errors_impl.InvalidArgumentError: Input 1 to node `gradients/layer_normalization/transpose_1_g
rad/transpose` with op Transpose must be a compile-time constant.

XLA compilation requires that operator arguments that represent shapes or dimensions be evaluated to concrete values at co
mpile time. This error means that a shape or dimension argument could not be evaluated at compile time, usually because th
e value of the argument depends on a parameter to the computation, on a variable, or on a stateful operation such as a ran
dom number generator.
```

TF1.15 Only (TF2 - D62672)

Test Plan: New tests

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jackh, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T48430

Differential Revision: https://phabricator.sourcevertex.net/D63383