Updating gather simplifier to support all valid gathers
Summary:
Gather simplifier converts gather ops into multi-slices.
Previously it only supported the trivial cases where we were slicing on dim zero, the data tensor was rank <= 2 and the indices tensor was rank <= 1.
This diff updates gather simplifier to support all valid gathers.
This includes any combination of the following:
  - Multiple slice dims
  - Multiple unsliced dims
  - Slice dims which don't start at zero
  - Arbitrary indices tensor shape (Gather subs out the sliced dims for this shape in the output)

Note that it is not possible to create a gather with multiple slice dims which are not all sequential, as it would be incompatible with the way output shape is calculated. Specifically, they way the indices shape is subbed in for the sliced dims.

TF1.15 Only

V2 diff: D61620

Differences from the v2 diff:
  - The addition of the MakeReduceHlo util which was already present in v2 but not v1 (Added in a531c6488294).
  - Reallocation of start_index_map in order to cast it to int64 in two places in gather_simplifier.cc.
  - Permutation passed to ShapeUtil::PermuteDimensions has to be inverted in v1 as the logic was inverted in v2.

Test Plan: Updated the existing gather simplifier tests to cover all of the relevant cases.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T50967

Differential Revision: https://phabricator.sourcevertex.net/D62360