Updating gather simplifier to support all valid gathers
Summary:
Gather simplifier converts gather ops into multi-slices.
Previously it only supported the trivial cases where we were slicing on dim zero, the data tensor was rank <= 2 and the indices tensor was rank <= 1.
This diff updates gather simplifier to support all valid gathers.
This includes any combination of the following:
  - Multiple slice dims
  - Multiple unsliced dims
  - Slice dims which don't start at zero
  - Arbitrary indices tensor shape (Gather subs out the sliced dims for this shape in the output)

Note that it is not possible to create a gather with multiple slice dims which are not all sequential, as it would be incompatible with the way output shape is calculated. Specifically, they way the indices shape is subbed in for the sliced dims.

TF2.5 Only

V1 diff: D62360

Test Plan: Updated the existing gather simplifier tests to cover all of the relevant cases.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, yanislavd, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, samuelh

Subscribers: babakk

Maniphest Tasks: T50967, T14037

Differential Revision: https://phabricator.sourcevertex.net/D61620