TF2: Simplify specifying the iterations when exporting pipelines for serving
Summary:
This change simplifies the way user specifies the number of iterations of the pipeline.
It replaces two previously used arguments: `pipeline_depth`, `iterations` with just a single one - `iterations`. That parameter
is now used as a pipeline's gradient accumulation count. Pipeline's `repeat_count` is now always set to the default value - 1,
as in the inference use cases, it does not matter that much as the gradient accumulation count.
That way user does not have to worry about optimal settings of both parameters and pipelines API is consistent with API for
exporting non-pipelined models which uses only `iterations` parameter.
Moreover, in Keras pipelined models, `repeat_count` is always set to 1 and user has the possibility of setting only gradient accumulation count
(denoted as `steps_per_execution`), so this aligns Keras and non-Keras implementations as well.

TF2.6 only

Ref T51876

Test Plan: Changed existing tests

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jacekr, jamiep, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jacekr, georgep

Subscribers: georgep, jakeh

Maniphest Tasks: T51876

Differential Revision: https://phabricator.sourcevertex.net/D64906