Start splitting poplar_compiler into smaller functions popit can use
Summary:
So that we can share code between popit and poplar compilers try split
shared parts into functions. This diff does this for the transform hlo
passes. It also changes compiler resources to only provide the resources
it uses to this function by creating a HloResources struct. This is mainly
just to ensure no one uses the graph in any of these passes.

This is a non functional change

TF1.15 only

lint

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jackh, vladimirm, babakk

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk

Subscribers: babakk

Maniphest Tasks: T57888

Differential Revision: https://phabricator.sourcevertex.net/D65077