TF2 - Removing IpuToMemorySchedulerAlgorithm
Summary:
Removing IpuToMemorySchedulerAlgorithm. This is in preparation of changing the IpuScheduleAlgorithm interface to use HloPoplarDataflowAnalysis, so it's no longer compatible with MemoryScheduleAlgorithm.

To do this I've moved tests that used HloMemorySchedule with an IpuScheduleAlgorithm to use IpuSchedule directly.

TF2.5 Only (due to TF1 merge failure)

TF1 - D65646

Test Plan: CI

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T9494

Differential Revision: https://phabricator.sourcevertex.net/D65611