Fix Mismatched Poplar/XLA shapes in program with static-multi-slice
Summary:
A `StaticMultiSlice` can be seen as a view-changing operation. However, it wasn't handled in the `PathTransform` which lead to wrong shapes during allocation. This diff adds a handler for it.

Fix T61219.

Test Plan: A new test based on the reproducer to check that static multi slices can now be allocated through.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, babakk

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk

Maniphest Tasks: T61219, T60337

Differential Revision: https://phabricator.sourcevertex.net/D66202