Improve handling of single computation liveness analysis.
Summary:
This change updates our liveness analysis to better handle single computations, which is the level at which the scheduler selection operates.

It now says that inplace parameters are alive for the duration of the computation - since a single computation doesn't provide enough context to determine the actual life span of the corresponding buffers.
Inplace outputs are now alive after the caller instruction. Previously they would be seen as being alive for the whole computation - since the creating instruction is out of the scope of the computation being analysed.
Memory estimates now include the costs of any subcomputations.

Test Plan: New tests

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, yanislavd, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, yanislavd

Maniphest Tasks: T9494

Differential Revision: https://phabricator.sourcevertex.net/D68034