Fix TF1 within_replica_tests
Summary:
`WithinReplicasTest.testOpShardOrder*` tests fail on TF1 as sharding information of non-scalar constants is being lost. This happens when they get replaced with a scalar constant + broadcast - sharding is only copied onto the broadcast but that can be elimated during `sinking of broadcast to user`. To fix this I've added a call to manually copy the original sharding onto the scalar constant.

This doesn't happen in TF2 due to a subtle difference in the xla graph generated. In TF2 when we ask for an array constant we get a scalar constant and broadcast, so avoid the optimisation that loses the sharding. In TF1 we get an actual array constant.

TF1.15 Only

Test Plan: CI

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T63279

Differential Revision: https://phabricator.sourcevertex.net/D68239