Preserve sharding when replacing array constants.
Summary:
When replacing an array constant with a scalar constant + broadcast the sharding is lost on the scalar constant. It gets added to the broadcast but that might be eliminated by a different optimisation.
This change manually adds sharding to the new scalar constant.

TF2 version of D68239

Test Plan: CI + new test.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T63279

Differential Revision: https://phabricator.sourcevertex.net/D68277