Fix rescaling for running mean
Summary:
When running with grouped schedule, the scale factor on the gradients needs to be adjusted to:
```
1 / ((accumulation_count - execution_counter) <= accumulation_count : (execution_counter + offset)
```
where offset is the number of stages from the stage which scales the gradient accumulation buffer to where the gradient is generated.
TF1.15 Only
Fix T63109

Test Plan: Added new tests, re-enabled failing tests.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, gauthamg, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, gauthamg

Maniphest Tasks: T63109

Differential Revision: https://phabricator.sourcevertex.net/D68892