Handling of preprocessing in export_single_step and export_pipeline
Summary:
This change introduces changes in model exporting API for both single step and pipelined models. It adds the ability to export preprocess function that is about to be run on CPU and that does produce the input for model running on IPU.
As an extra, it provides interface for providing path to checkpoint which gets restored just before saving the graph, so that saved variables' values can be retrieved back and saved properly.

The patch provides both code for new API, documentation, tests and examples.

Closes T64949

TF1.15 Only

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, piotrc, bartlomiejw, #tech_docs, samuelh, jackh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, piotrc, bartlomiejw, jackh

Subscribers: georgep, jayniep, piotrc, bartlomiejw

Maniphest Tasks: T64949

Differential Revision: https://phabricator.sourcevertex.net/D70126