Inserting copies when we can't otherwise inplace
Summary:
If an op does not support non-inplacing but it cannot be inplaced without creating a cycle, insert a copy to allow it to be inplaced.
This is a the best option when inplacing calls because other parameters which do not require copies would otherwise miss out on being inplace.

operands_to_copy is changed to a map of instruction -> vector of operand indices, which makes it more straightforward to insert into.

Test Plan: Added a test which tests that pipeline stages with a parameter in common are still inplaced by inserting a copy.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm

Maniphest Tasks: T63089

Differential Revision: https://phabricator.sourcevertex.net/D70783