Adding elementwise preapply pass
Summary:
Adds a new pass to deal with a particular type of optimisation which doesn't fit well into the algebraic simplifier.

The optimisation identifies "uniform" elementwise ops applied to specfic target ops. "Uniform" means every element of the target is modified in the same way, i.e. every input except the target is a broadcasted scalar. For certain targets (e.g. broadcast), these elementwise ops can be pre-applied to smaller inputs of the target rather than the larger output.

The target op handled in this diff is onehot. The elementwise op is preapplied to the on and off values.

The pattern of doing a onehot with 0/1 for on/off,  then multiplying and adding to transform the output into the desired range is seen in tf2 resnet. In the resnet case, the add and multiply get further optimised with constant folding.

Test Plan: Added a new test file.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T65384, T63089

Differential Revision: https://phabricator.sourcevertex.net/D71126