Consider copying slice outputs to save memory in a scheduled graph
Summary:
This pass starts by traversing the scheduled graph. It counts the number of branches through which each buffer flows in and out of an instruction, essentially keeping a count of the number of different paths in which each buffer is live. An instruction is copied if all conditions below are met:

- The instruction is `HloOpcode::kSlice`.
- Copying the instruction would result in reduction of live memory.
- The maximum from the number of live paths for each buffer that flows in the instruction is 1 (this guarantees the buffers will be killed after the instruction is copied).
- The parent computation for that instruction is unique (i.e. not a clone that `xla::ModuleFlatten` created).

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, vladimirm, babakk

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk

Maniphest Tasks: T43867

Differential Revision: https://phabricator.sourcevertex.net/D67035