Check for size = 1 instead of IsScalar in ElementwisePreapply
Summary:
This is more general, doesn't affect the "uniformity" of the operand, and should therefore allow for the pass to be applied in more cases.
Fix T67670.

Test Plan: Added tests for when the output size is 1. Also added tests for when elementwise operands have different types, since this was an issue in another test.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alfiee

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alfiee

Maniphest Tasks: T67670

Differential Revision: https://phabricator.sourcevertex.net/D73213